Have default client protocol
