+++ tests/test_scream.py
@@ -27,7 +27,7 @@ class TestScream_01_Scream(_TestOOC):
         # C0 (non-staff) screams
         self.c0.ooc('/scream Hi')
         self.c0.assert_ooc('You screamed "Hi".', ooc_over=True)
-        self.c0.assert_ic('Hi', cid=0, pos=self.c0.pos, showname=self.c0.showname, over=True)
+        self.c0.assert_ic('Hi', char_id=0, pos=self.c0.pos, showname=self.c0.showname, over=True)
         self.c1.assert_ooc('(X) {} [{}] screamed "Hi" ({}).'
                            .format(self.c0_dname, 0, 0), ooc_over=True)
         self.c1.assert_ic('Hi', over=True)
@@ -48,7 +48,7 @@ class TestScream_01_Scream(_TestOOC):
                            .format(self.c3_dname, 3, 5), over=True)
         self.c2.assert_no_ic()  # C2 is not in scream range
         self.c3.assert_ooc('You screamed "Hu".', ooc_over=True)
-        self.c3.assert_ic('Hu', cid=3, pos=self.c3.pos, showname=self.c3.showname, over=True)
+        self.c3.assert_ic('Hu', char_id=3, pos=self.c3.pos, showname=self.c3.showname, over=True)
 
     def test_03_staffscream(self):
         """
@@ -60,7 +60,7 @@ class TestScream_01_Scream(_TestOOC):
         self.c0.assert_ooc('You heard {} scream nearby.'.format(self.c1_dname), ooc_over=True)
         self.c0.assert_ic('Ha', showname=self.c1.showname, over=True)
         self.c1.assert_ooc('You screamed "Ha".', ooc_over=True)
-        self.c1.assert_ic('Ha', cid=1, pos=self.c1.pos, showname=self.c1.showname, over=True)
+        self.c1.assert_ic('Ha', char_id=1, pos=self.c1.pos, showname=self.c1.showname, over=True)
         self.c2.assert_ooc('(X) {} [{}] screamed "Ha" ({}).'.format(self.c1_dname, 1, 0),
                            over=True)
         self.c2.assert_no_ic()  # C2 is not in scream range
@@ -73,7 +73,7 @@ class TestScream_01_Scream(_TestOOC):
                            over=True)
         self.c1.assert_no_ic()  # C1 is not in scream range
         self.c2.assert_ooc('You screamed "He".', ooc_over=True)
-        self.c2.assert_ic('He', cid=2, pos=self.c2.pos, showname=self.c2.showname, over=True)
+        self.c2.assert_ic('He', char_id=2, pos=self.c2.pos, showname=self.c2.showname, over=True)
         self.c3.assert_no_packets()  # C3 is not in scream range
 
     def test_04_screamwhilemutedglobal(self):
@@ -105,5 +105,5 @@ class TestScream_01_Scream(_TestOOC):
         self.c2.assert_ooc('(X) {} [{}] screamed "and let it all out" ({}).'
                            .format(self.c3_dname, 3, 5), over=True)
         self.c3.assert_ooc('You screamed "and let it all out".', ooc_over=True)
-        self.c3.assert_ic('and let it all out', cid=3, pos=self.c3.pos, showname=self.c3.showname,
-                          over=True)
+        self.c3.assert_ic('and let it all out', char_id=3, pos=self.c3.pos,
+                          showname=self.c3.showname, over=True)
