+++ tests/test_whisper.py
@@ -177,11 +177,11 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
-                                 showname=sender.showname, over=True)
+                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos,
+                                 char_id=sender.char_id, showname=sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
-                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
-                                    showname=sender.showname, over=True)
+                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos,
+                                    char_id=sender.char_id, showname=sender.showname, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
@@ -208,11 +208,11 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                                 .format(sender.displayname, recipient.displayname))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
+                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, char_id=sender.char_id,
                                  showname=sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
-                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
-                                    showname=sender.showname, over=True)
+                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos,
+                                    char_id=sender.char_id, showname=sender.showname, over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
                 self.c4.assert_ooc(nonstaff_ooc, over=True)
@@ -237,11 +237,11 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
+                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, char_id=sender.char_id,
                                  showname=sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
-                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
-                                    showname=sender.showname, over=True)
+                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos,
+                                    char_id=sender.char_id, showname=sender.showname, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
                 self.c4.assert_no_packets()
@@ -304,7 +304,7 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                 client.ooc('/whisper {} {}'.format(identifier, message))
                 client.assert_ooc(ooc, ooc_over=True)
                 client.assert_ic(message, folder="<NOCHAR>", pos=client.pos,
-                                 showname=client.showname, cid=client.char_id, over=True)
+                                 showname=client.showname, char_id=client.char_id, over=True)
 
     def test_06_whispersneakedtosneakedparty(self):
         """
@@ -352,11 +352,11 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
+                sender.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, char_id=sender.char_id,
                                  showname=sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
-                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
-                                    showname=sender.showname, over=True)
+                recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos,
+                                    char_id=sender.char_id, showname=sender.showname, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_no_packets()
                 self.c3.assert_no_packets()
@@ -425,11 +425,11 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos=sender.pos,
-                                 cid=sender.char_id, showname=sender.showname,
+                                 char_id=sender.char_id, showname=sender.showname,
                                  allow_partial_match=True, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos=sender.pos,
-                                    cid=sender.char_id, showname=sender.showname,
+                                    char_id=sender.char_id, showname=sender.showname,
                                     allow_partial_match=True, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
@@ -458,10 +458,10 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
             sender.ooc('/whisper {} {}'.format(identifier, message))
             sender.assert_ooc(sent_ooc, ooc_over=True)
             sender.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos=sender.pos,
-                             cid=sender.char_id, showname=sender.showname,
+                             char_id=sender.char_id, showname=sender.showname,
                              allow_partial_match=True, over=True)
             recipient.assert_ooc(recipient_ooc, ooc_over=True)
-            recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, cid=sender.char_id,
+            recipient.assert_ic(message, folder='<NOCHAR>', pos=sender.pos, char_id=sender.char_id,
                                 showname=sender.showname, over=True)
             self.c2.assert_ooc(staff_ooc, over=True)
             self.c3.assert_no_packets()
@@ -525,7 +525,7 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 client.ooc('/whisper {} {}'.format(identifier, message))
                 client.assert_ooc(ooc, ooc_over=True)
                 client.assert_ic('(Your ears are ringing)', folder="<NOCHAR>", pos=client.pos,
-                                 showname=client.showname, cid=client.char_id,
+                                 showname=client.showname, char_id=client.char_id,
                                  allow_partial_match=True, over=True)
 
     def test_05_whispersneakedtosneakedparty(self):
@@ -576,11 +576,11 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos=sender.pos,
-                                 cid=sender.char_id, showname=sender.showname,
+                                 char_id=sender.char_id, showname=sender.showname,
                                  allow_partial_match=True, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos=sender.pos,
-                                    cid=sender.char_id, showname=sender.showname,
+                                    char_id=sender.char_id, showname=sender.showname,
                                     allow_partial_match=True, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_no_packets()
